+++ scripts/regenerate-images.sh
@@ -8,6 +8,7 @@ stack exec waterfall-cad-examples --  --offset --dark-mode-svg images/offset.svg
 stack exec waterfall-cad-examples --  --text --content "Waterfall CAD" --dark-mode-svg images/text.svg --font images/fonts/varela/VarelaRound-Regular.ttf 
 stack exec waterfall-cad-examples --  --bound --dark-mode-svg images/bounding-boxes.svg --height 200
 stack exec waterfall-cad-examples --  --loft --dark-mode-svg images/loft.svg
+stack exec waterfall-cad-examples --  --2d-booleans --dark-mode-svg images/2d-booleans.svg --height 200
 
 stack exec waterfall-cad-examples --  --csg --glb images/models/csg.glb
 stack exec waterfall-cad-examples --  --gear --glb images/models/gear.glb
@@ -19,3 +20,4 @@ stack exec waterfall-cad-examples --  --bound --glb images/models/bounding-boxes
 stack exec waterfall-cad-examples --  --loft --glb images/models/loft.glb
 stack exec waterfall-cad-examples --  --fillet --glb images/models/fillet.glb
 stack exec waterfall-cad-examples --  --prism --glb images/models/prism.glb
+stack exec waterfall-cad-examples --  --2d-booleans --glb images/models/2d-booleans.glb